Finish frontend; refactor backend

+++ api/routes/user/user.controller.js
@@ -1,31 +1,6 @@
 const { PrismaClient } = require('@prisma/client')
+const { handleError, respondWithResult, userExists } = require('../../services')
 const prisma = new PrismaClient()
-
-// ==================================================================
-// Helpers
-// ==================================================================
-const handleError = (res, statusCode = 500) => error => {
-  console.log(error.message)
-  res.status(statusCode).send(error.message || error)
-}
-
-const respondWithResult = (res, statusCode = 200) => entity => {
-  if (entity) {
-    res.status(statusCode).json(entity)
-  } else {
-    handleError(res, 500)({ message: 'Entity not found' })
-  }
-}
-
-const userExists = res => async id => {
-  const user = await prisma.user.findUnique({
-    where: { id },
-  })
-
-  if (!user) {
-    handleError(res, 404)({ message: 'User does not exist' })
-  }
-}
 
 // ==================================================================
 // Controls
@@ -46,21 +21,15 @@ exports.create = async function (req, res) {
   }
 }
 
-// GET to list all users
-exports.list = async function (req, res) {
-  try {
-    const users = await prisma.user.findMany()
-    return respondWithResult(res, 200)(users)
-  } catch (error) {
-    return handleError(res, 500)(error)
-  }
-}
-
 // GET to show a user
 exports.show = async function (req, res) {
   const id = parseInt(req.params.id, 10)
 
   try {
+    if (!await userExists(id)) {
+      return handleError(res, 404)({ message: 'User does not exist' })
+    }
+
     const user = await prisma.user.findUnique({
       where: { id },
     })
@@ -81,7 +50,9 @@ exports.update = async function (req, res) {
   const id = parseInt(req.params.id, 10)
 
   try {
-    await userExists(res)(id)
+    if (!await userExists(id)) {
+      return handleError(res, 404)({ message: 'User does not exist' })
+    }
 
     const user = await prisma.user.update({
       where: { id },
@@ -99,10 +70,13 @@ exports.destroy = async function (req, res) {
   const id = parseInt(req.params.id, 10)
 
   try {
-    await userExists(res)(id)
+    if (!await userExists(id)) {
+      return handleError(res, 404)({ message: 'User does not exist' })
+    }
+
     await prisma.user.delete({ where: { id } })
-    return res.sendStatus(204)
+    return respondWithResult(res, 204)()
   } catch (error) {
-    return handleError(res, 404)({ message: 'User does not exist' })
+    return handleError(res, 500)(error)
   }
 }
